refactor: move configure_system_for_boot into arch module
Split `configure_system_for_boot` into 2 arch specific functions
and move it into arch module.
